Logre mostrar un dialog
