+++ src/app/dialog-especialistas/dialog-especialistas.component.html
@@ -4,9 +4,9 @@
 
 <div mat-dialog-content>
     <p>What's your favorite animal?</p>
-    <mat-form-field>
+    <!-- <mat-form-field>
       <input matInput [(ngModel)]="data.animal">
-    </mat-form-field>
+    </mat-form-field> -->
   </div>
   <div mat-dialog-actions>
     <button mat-button (click)="onNoClick()">No Thanks</button>
